refactor api endpoints
